Add SAMS server implementation for match and ranking queries
This commit introduces a new server-only module `sams.server.ts` that encapsulates the logic for fetching SAMS match data and rankings, ensuring it is excluded from client bundles. The existing `sams.ts` file has been refactored to export server function wrappers, enhancing code organization and clarity. Additionally, test imports have been updated to reflect the new structure.
