+++ app/src/server/functions/sams.test.ts
@@ -2,14 +2,13 @@ import { InvokeCommand, LambdaClient } from "@aws-sdk/client-lambda";
 import { mockClient } from "aws-sdk-client-mock";
 import { beforeEach, describe, expect, it } from "vite-plus/test";
 import {
-  buildLiveMatchesFromRaw,
-  createSamsMatchesCacheKey,
-  invokeSamsLambdaAsync,
-  resolveEffectiveSamsSportsclubUuids,
-  resolveClubLogoUrl,
-  triggerSamsClubsSyncFn,
-  triggerSamsTeamsSyncFn,
-} from "./sams";
+    buildLiveMatchesFromRaw,
+    createSamsMatchesCacheKey,
+    invokeSamsLambdaAsync,
+    resolveEffectiveSamsSportsclubUuids,
+    resolveClubLogoUrl,
+} from "./sams.server";
+import { triggerSamsClubsSyncFn, triggerSamsTeamsSyncFn } from "./sams";
 
 describe("resolveClubLogoUrl", () => {
   const CF = "https://cdn.example.com";
